fastapi-crud/requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
pydantic[email]
